ensure a form has a touchpoint before trying to delete it

+++ db/migrate/20191024232545_delete_custom_forms.rb
@@ -4,8 +4,10 @@ class DeleteCustomForms < ActiveRecord::Migration[5.2]
       next if form.kind == "custom"
       next if form.touchpoint && form.touchpoint.submissions.present?
 
-      puts "destroying Touchpoint #{form.touchpoint.id} #{form.touchpoint.name}"
-      form.touchpoint.destroy
+      if form.touchpoint
+        puts "destroying Touchpoint #{form.touchpoint.id} #{form.touchpoint.name}"
+        form.touchpoint.destroy
+      end
       puts "destroying Form #{form.id} #{form.name} #{form.title}"
       form.destroy
     end
